Mark table text for display in landscape environment

Important: this function will not make your table render in landscape mode in a stand alone TeX document. This function doesn't actually make the table "landscape". It simply adds to the class attribute so that functions down the line can look for this tag and wrap the table in landscape environment. The landscaping is entirely the responsibility of some other function, just not this one.

as_lscape(x)

is_lscape(x)

Arguments

x

output from either stable() or stable_long()

Value

For as_lscape(), the input is returned, but with stable_lscape added to the class attribute. For is_lscape(), a logical value is returned, indicating whether or not the object inherits from stable_lscape.

See also

Examples

out <- stable(stdata())
is_lscape(out)
#> [1] FALSE

land <- as_lscape(out)
is_lscape(land)
#> [1] TRUE

st_wrap(as_lscape(out))
#> ```{=latex}
#> \begin{landscape}
#> \begin{table}[H]
#> \centering
#> \setlength{\tabcolsep}{5pt} 
#> \begin{threeparttable}
#> \renewcommand{\arraystretch}{1.3}
#> \begin{tabular}[h]{lllllllll}
#> \hline
#> STUDY & DOSE & FORM & N & WT & CRCL & AGE & ALB & SCR \\
#> \hline
#> 12-DEMO-001 & 100 mg & tablet & 80 & 71.4 & 104 & 33.7 & 4.20 & 1.06 \\
#> 12-DEMO-001 & 150 mg & capsule & 16 & 89.4 & 122 & 24.4 & 4.63 & 1.12 \\
#> 12-DEMO-001 & 150 mg & tablet & 48 & 81.7 & 104 & 34.4 & 3.83 & 0.910 \\
#> 12-DEMO-001 & 150 mg & troche & 16 & 94.0 & 93.2 & 27.4 & 4.94 & 1.25 \\
#> 12-DEMO-001 & 200 mg & tablet & 64 & 67.9 & 100 & 27.5 & 4.25 & 1.10 \\
#> 12-DEMO-001 & 200 mg & troche & 16 & 76.6 & 99.2 & 22.8 & 4.54 & 1.15 \\
#> 12-DEMO-002 & 100 mg & capsule & 36 & 61.3 & 113 & 38.3 & 4.04 & 1.28 \\
#> 12-DEMO-002 & 100 mg & tablet & 324 & 77.6 & 106 & 29.9 & 4.31 & 0.981 \\
#> 12-DEMO-002 & 50 mg & capsule & 36 & 74.1 & 112 & 37.1 & 4.44 & 0.900 \\
#> 12-DEMO-002 & 50 mg & tablet & 324 & 71.2 & 106 & 34.1 & 4.63 & 0.868 \\
#> 12-DEMO-002 & 75 mg & capsule & 36 & 72.4 & 105 & 38.2 & 3.89 & 0.900 \\
#> 12-DEMO-002 & 75 mg & tablet & 288 & 71.6 & 98.9 & 34.2 & 4.49 & 0.991 \\
#> 12-DEMO-002 & 75 mg & troche & 36 & 73.6 & 103 & 49.2 & 4.52 & 0.930 \\
#> \hline
#> \end{tabular}
#> \end{threeparttable}
#> \end{table}
#> \end{landscape}
#> ```
#> 
#>