Class: pmtable

Objects with class pmtable are return from the pmtable-provided data summary functions:

pt_cont_wide()

pt_cont_long()

pt_cat_wide()

pt_cat_long()

pt_data_inventory()

Details

The object is a list, with an item called data, which is the summarized data set, as well as other items that are arguments for stable() or stable_long().

Examples

tab <- pt_cont_long(pmt_first, cols = "WT,AGE,SCR")

str(tab)
#> List of 5
#>  $ data     : tibble [3 × 6] (S3: tbl_df/tbl/data.frame)
#>   ..$ Variable : chr [1:3] "WT" "AGE" "SCR"
#>   ..$ n        : chr [1:3] "157" "160" "160"
#>   ..$ Mean     : chr [1:3] "70.7" "33.7" "1.36"
#>   ..$ Median   : chr [1:3] "70.0" "33.4" "1.04"
#>   ..$ SD       : chr [1:3] "12.8" "8.83" "0.986"
#>   ..$ Min / Max: chr [1:3] "43.6 / 97.2" "18.9 / 49.5" "0.710 / 5.59"
#>  $ align    :List of 5
#>   ..$ complete: NULL
#>   ..$ default : chr "c"
#>   ..$ update  : list()
#>   ..$ coltype : chr "p"
#>   ..$ outer   : chr "lr"
#>   ..- attr(*, "class")= chr "aligncol"
#>  $ panel    :List of 12
#>   ..$ col        : chr ""
#>   ..$ prefix     : chr ""
#>   ..$ prefix_name: logi FALSE
#>   ..$ prefix_skip: NULL
#>   ..$ null       : logi TRUE
#>   ..$ dup_err    : logi TRUE
#>   ..$ bold       : logi TRUE
#>   ..$ it         : logi FALSE
#>   ..$ skip       : chr ".panel.skip."
#>   ..$ hline      : logi TRUE
#>   ..$ jut        : num 0
#>   ..$ nopagebreak: logi TRUE
#>   ..- attr(*, "class")= chr "rowpanel"
#>  $ bold_cols: logi TRUE
#>  $ notes    : chr [1:3] "n: number of records summarized" "SD: standard deviation" "Min: minimum; Max: maximum"
#>  - attr(*, "class")= chr [1:2] "pmtable" "list"

methods(class = "pmtable")
#> [1] as_stable   st_as_image st_new      stable      stable_long
#> see '?methods' for accessing help and source code