Create continuous summary data frame

Create continuous summary data frame

cont_table_data(
  data,
  cols,
  by = ".total",
  panel = by,
  wide = FALSE,
  all_name = "all",
  digits = new_digits(),
  id_col = "ID",
  na_fill = "--",
  fun = cont_long_fun
)

Arguments

data

the data frame to summarize; the user should filter or subset so that data contains exactly the records to be summarized; pmtables will not add or remove rows prior to summarizing data

cols

the columns to summarize; may be character vector or quosure

by

grouping variable name

panel

paneling variable name

wide

logical; if TRUE, output will be wide; output will be long otherwise

all_name

label for full data summary

digits

named list specifying digits argument for digit_fun

id_col

the ID column name

na_fill

value to fill with when all values in the summary are missing

fun

continuous data summary function