Configure table notes

Configure table notes

noteconf(
  width = 0.8,
  type = c("tpt", "minipage"),
  hline = c("top", "bottom", "both", "none"),
  table_skip = 0.67,
  note_skip = 0.02,
  linespread = 1.1,
  hline_pt = 0.4,
  note_sp = 0.1
)

Arguments

width
the width (as fraction of linewidth) is minipage notes

type
put table notes in either the tablenotes section of threeparttable (tpt) or in a minipage below the table (minipage)

hline
where to include horizintal lines

table_skip
vertical space (cm) between the last row of the table and the hline (if any) at the top of the table notes

note_skip
vertical space (cm) between the top hline of the table notes and the first note line

linespread
spacing between note lines for minipage notes

hline_pt
the point size for hlines

note_sp
spacing between note lines for tpt notes