Set global plot options

Set global plot options

pt_options(
  id_col = "ID",
  dv_col = "DV",
  bq_col = "BQL",
  units = NULL,
  table = NULL,
  fun.cont.long = pmtables:::df_sum_2,
  fun.cont.wide = pmtables:::str_sum_2,
  digits = new_digits()
)

pt_opts

opts

Arguments

id_col
ID column name

dv_col
DV column name

bq_col
BQL column name

units
named list of units

table
named list of column name transformations

Format

An object of class pt_opts (inherits from environment) of length 15.

An object of class pt_opts (inherits from environment) of length 15.

Details

pt_opts is the options environment.

Global options can be set in the environment with pt_opts$set(name = value). There is also a .list argument to pt_opts$set that allows you to pass in a named list of options to set.

Values can be extracted with pt_opts$get("name").

Because it is an environment, the $ operator can also be used to get and set values (see examples).

Other methods in the environment include pt_opts$as.list(), pt_opts$mget(), pt_opts$reset(). pt_opts$self refers to the environment itself. A list of default settings can be obtained with pt$defaults. Methods for the pt_opts object include: print.pt_opts, as.list.pt_opts, $<-.pt_opts.

Examples

pt_opts$set(id_col = "USUBJID")

pt_opts$id_col <- "ID"

pt_opts$bq_col

#> [1] "BLQ"

pt_opts$reset()

x <- pt_opts$as.list()

if (FALSE) {
 defs <- pt_opts$defaults
 defs$dv_col <- "dv"
 pt_opts$set(.list = defs)
}