Create panel object

Objects may also be coerced with as.panel()

rowpanel(
  col = NULL,
  prefix = "",
  skip = ".panel.skip.",
  prefix_name = FALSE,
  prefix_skip = NULL,
  duplicates_ok = FALSE,
  bold = TRUE,
  it = FALSE,
  hline = TRUE,
  jut = 0,
  nopagebreak = TRUE
)

is.rowpanel(x)

Arguments

col

name of column to be used for creating panels

prefix

to be added to each panel title

skip

regular expression for panels to skip

prefix_name

logical; if TRUE, then the prefix will be derived by the name associated with col input to the function

prefix_skip

a regular expression for identifying panel titles where the prefix won't be applied

duplicates_ok

if FALSE, an error is generated if more than one panel will have the same header

bold

render panel title in bold font face

it

render panel title in italic font face

hline

logical indicating whether or not to draw an hline above the panel row; the first panel row never receives an hline

jut

amount (in TeX ex units) by which the panel headers are outdented from the first column in the main table and header. Consider using a value of 2 for a clear offset.

nopagebreak

if TRUE (the default) then the page will not break immediately after a panel title when creating a longtable; this argument does nothing if you are not generating a longtable. Set to FALSE to prevent the nopagebreak command (implemented as *) in the longtable output.

x

an object to test

See also