Table control parameters

Table control parameters

st_control(...)

# S3 method for default
st_control(
  row_space = 1.4,
  col_space = 5,
  lt_row_space = 0.3,
  fontsize = "normalsize",
  header_row = 0.56,
  note_type = c("tpt", "minipage"),
  note_hline = c("top", "bottom", "both", "none"),
  note_hline_pt = 0.4,
  note_table_skip = 0.67,
  note_skip = 0.02,
  note_sanitize = TRUE,
  note_escape = "_",
  ...
)

# S3 method for stobject
st_control(x, ...)

Arguments

...
used only in the generic

row_space
multiplier to increase or decrease padding between table rows; for example, row_space = 1.4 means increase by 40%

col_space
padding between columns; unit is pt

lt_row_space
extra space between rows of longtable; unit is em

fontsize
font size

header_row
amount of space that will be removed between rows of the table header; unit is em; this must be a positive number

note_type
tpt puts notes in the third part of threeparttable and minipage puts notes in a minipage below the table

note_hline
where to place hlines when minipage notes are used

note_hline_pt
pointsize for minipage note hline

note_table_skip
vertical space between the bottom of the table and the start of minipage notes; unit is cm

note_skip
vertical space after the top hline where the minipage notes start

note_sanitize
if TRUE, notes will be sanitized

note_escape
a vector of characters to sanitize

x
an stobject