Convert existing note vector into a single string

Convert existing note vector into a single string

st_notes_str(x, collapse = "; ")

Arguments

x

an stobject.

collapse

a character string to separate notes which are pasted together when flattening or appending; this should usually end in a single space (see default).

Value

An updated object with class stobject, which can be piped to other functions.