Create tabular output from an R data frame

Create tabular output from an R data frame

stable(
  data,
  align = cols_left(),
  panel = rowpanel(col = NULL),
  units = NULL,
  clear_reps = NULL,
  rm_dups = NULL,
  span = NULL,
  span_split = NULL,
  notes = NULL,
  hline_at = NULL,
  hline_from = NULL,
  sumrows = NULL,
  bold_cols = NULL,
  col_rename = NULL,
  col_blank = NULL,
  col_replace = NULL,
  row_space = 1.4,
  col_space = 5,
  fontsize = NULL,
  prime_fun = tab_prime,
  escape_fun = tab_escape,
  note_config = noteconf(type = "tpt"),
  r_file = NULL,
  r_file_label = getOption("r.file.label", "Source code:"),
  output_file = NULL,
  output_file_label = getOption("out.file.label", "Source file: ")
)

Arguments

data
a data.frame to convert to tabular table; see also st_new()

align
an alignment object created by cols_align(), cols_left(), cols_center(), or cols_right(); see also st_align()

panel
character column name to use to section the table; sections will be created from unique values of data[[panel]]; see also st_panel()

units
a named list with unit information; names should correspond to columns in the data frame

clear_reps
character vector of column names where duplicate values will be made blank (overwritten with ""); ; see also st_clear_reps()

rm_dups
deprected; use clear_reps

span
a list of objects created with colgroup(); ; see also st_span()

span_split
not implemented at this time; ; see also st_span_split()

notes
a character vector of notes to include at the foot of the table; use r_file and output_file for source code and output file annotations; see also st_notes()

hline_at
logical or integer vector specifying rows above which an \hline will be placed; see also st_hline()

hline_from
a character column name from which to separate the table with \hline; non-duplicated values of hline_from will be used to create the split; see also st_hline()

sumrows
an object created with sumrow(); identifies summary rows and adds styling; see also st_sumrow()

bold_cols
if TRUE, table column names are rendered with bold font

col_rename
a name = value character vector to translate column names to table names; ; see also st_rename()

col_blank
a character vector of column names that will not be printed in the table header; see also st_blank()

col_replace
a character vector with the same length as the number of output table columns; use this to completely replace the names (as opposed to one by on editing with col_rename)

row_space
relative increase or decrease spacing between rows; use row_space > <default> to increase; ; see also st_space()

col_space
absolute column spacing amount (pt); see also st_space()

fontsize
for the table (e.g. normalsize, small, scriptsize, etc)

prime_fun
function to prime the data frame to be converted to tabular

escape_fun
a function passed to prime_fun that will sanitize column data

note_config
a noteconf() object used to configure how table notes are displayed; ; see also st_noteconf()

r_file
the name of the R file containg code to generate the table; the file name will be included in the notes in the table footer; ; see also st_files()

r_file_label
prefix text for r_file note

output_file
the name of the output file where the table text will be saved; the file name will be included in the notes in the table footer; see also st_files()

output_file_label
prefix text for output_file note