Clear repeated values in a column

Clear repeated values in a column

tab_clear_reps(
  data = NULL,
  clear_reps = NULL,
  panel = rowpanel(NULL),
  clear_grouped_reps = NULL,
  ...
)

Arguments

data

a data.frame to convert to tabular table; the user should filter or subset so that data contains exactly the rows (and columns) to be processed; pmtables will not add or remove rows prior to processing data; see also st_new()

clear_reps

character vector of column names where duplicate values will be made blank (overwritten with ""); see also st_clear_reps()

panel

character column name to use to section the table; sections will be created from unique values of data[[panel]]; see also st_panel()

clear_grouped_reps

passed to clear_grouped_values() as cols

...

not used

See also