Adjustments for row padding, column padding, and font size

Adjustments for row padding, column padding, and font size

tab_size(
  row = 1.3,
  col = 5,
  font = NULL,
  header_row = -0.4 * row,
  span_title_row = header_row,
  lt_row = 0
)

Arguments

row

relative increase or decrease spacing between rows; use row_space > <default> to increase; this parameter applies to both long tables ans tabular tables; see also st_space()

col

absolute column spacing amount (pt); this parameter applies to both long tables and tabular tables; see also st_space()

font

for the table (e.g. normalsize, small, scriptsize, etc)

header_row

extra (or less) space between rows in the header section of the table (column names and units); this should be a negative number if you wish to compress header rows; the default is -0.4*row; recommend fixing to -0.2 when row is 1 and smaller if row is less than 1

span_title_row

extra (or less) space between rows of the span titles when they are broken across multiple lines; this is equivalent to header_row (which applies to table column titles), but this applies to spanner titles

lt_row

extra row space for longtables; this is only needed for extra control over row height; this can be a positive or negative number and a value of 0 indicates to neither add or subtract row space