Create groups of columns with spanners

Create groups of columns with spanners

tab_spanners(
  data,
  cols = NULL,
  span = NULL,
  span_split = NULL,
  span_title_break = "...",
  sizes = tab_size(),
  ...
)

Arguments

data

a data.frame to convert to tabular table; the user should filter or subset so that data contains exactly the rows (and columns) to be processed; pmtables will not add or remove rows prior to processing data; see also st_new()

cols

a character vector of column names

span

a list of objects created with colgroup(); ; see also st_span()

span_split

a colsplit object ; ; see also st_span_split()

span_title_break

a character sequence indicating where to split the title across multiple lines

sizes

an object returned from tab_size()

...

not used

See also