pmtables: Tables for Pharmacometrics.

Summarize data sets and create publication-quality tables for inclusion in 'tex' documents.

Creating general tables

To create a table with a single function call, use stable() for single page tables and stable_long() for longtable

Alternatively, use st_new() to create a s-table object and pipe to other st-functions; see the documentation links in stable() to discover these functions; finally, create an stable object with st_make()

Table data and grooming

The functions tab_prime() and triage_data() are called to get the data frame ready to be in tabular format (see make_tabular()

The function tab_escape() is used to escape % and _ in the table (sanitization); use options(pmtables.escape = ...) to change the set of escape characters or set that option to NULL to prevent any sanitization

Summarizing and creating tables

Use pt_cont_wide() to create continuous data summary in wide format

Use pt_cont_long() to create continuous data summary in long format

Use pt_cat_wide() to create discrete data summaries in wide format

Use pt_cat_long() to create discrete data summaries in long format

Use pt_data_inventory() to create a table summarizing observations and individuals in a data set

All of the above functions return an object with class pmtable, which is a list; you can access the summarized data by looking at the data slot; otherwise, pass the pmtable object to as_stable() in order to create the table

You can configure the digits in these summaries with new_digits()

Helper functions for working with data frames

Make text bold with tex_bold() or italics with tex_it()

Clear repeated values in a column with clear_grouped_values() and tab_clear_reps()

Use tab_edit() to do find / replace on your data frame

Use df_grep_rows() and df_grepl_rows() to find rows that match a given pattern

Use sig() or digit1() to control the number of digits that are displayed; both these functions return character

Functions to align columns

When you want latex to figure out column widths: cols_left(), cols_right(), cols_center(), cols_align()

When you want to fix the width of a specific column in the table: col_ragged(), col_fixed()

Functions to modify or configure a table

Alter column names: tab_cols()

Partition a table into panels: as.panel() and rowpanel()

Create groups of columns with column spanners: colgroup() and colsplit()

Configure row spacing, column spacing, or font size: tab_size()

Configure the appearance of table notes: noteconf()

Identify rows that are "summary" rows: sumrow(); these rows can get automatically styled with a label (optionally in bold font) and a horizontal line above

Preview tables

Use st_preview() to send s-table output to texPreview

Use st2doc() to render a pdf file with one or more tables

Use pt_wrap() to wrap s-table output in a table environment and optionally send the output to stdout(); this is helpful when rendering tables in Rmarkdown documents

Save s-tables

Use stable_save() to write an s-table object to file

Data sets

analysis1 - a NMTRAN-style data set; the basis for most other example data sets

pmt_pk - this is analysis1, but with only PK observation records

pmt_obs - this is analysis1, but with only observations

pmt_first - this is analysis1 but with only the first record from each individual

pmt_summarized - an example data set that has been summarized