Create a data inventory table

Output columns include counts for subjects, observations, BQL observations, and missing observations and percentage of observations that are BQL.

pt_data_inventory(
  data,
  outer = ".total",
  inner = outer,
  inner_summary = TRUE,
  drop_miss = FALSE,
  stacked = FALSE,
  table = NULL,
  all_name = "all",
  ...
)

Arguments

data
the data frame to summarize

outer
the outer grouping variable; may be character or quosure

inner
the inner grouping variable; may be character or quosure

inner_summary
if TRUE, then a summary of the inner variable will be provided

drop_miss
if TRUE, then MISS will be dropped, but only when all MISS values are equal to zero

stacked
if TRUE, then independent summaries are created by outer and included in a single table (see examples)

table
a named list to use for renaming columns (see details and examples)

all_name
a name to use for the complete data summary

...
other arguments passed to data_inventory_chunk

Details

The summary function is expecting certain columns to be named in a certain way. This can be modified to suit your need by passing the following arguments: dv_col (for observations), bq_col (for BQL observations), and id_col (for ID). See the data_inventory_chunk help topic for a description of these columns.

Examples

data <- pmtables:::data("id")

ans <- pt_data_inventory(data, outer = .cols("Renal function" = RFf))