Convert stable text to a standalone snippet

Convert stable text to a standalone snippet

st_to_standalone(
  text,
  stem,
  dir,
  font = c("helvetica", "roboto", "utopia"),
  command = "latex",
  textwidth = 6.5,
  border = "0.2cm 1cm",
  ntex = 1,
  ltversion = getOption("pmtables.image.ltversion", 4.13)
)

Arguments

text

character vector of table text.

stem

used to build intermediate and output file names.

dir

directory for building the pdf file.

font

the font to use; alternative values include roboto and utopia; passed to st_to_standalone().

command

pass pdflatex when building a pdf file or latex when building png.

textwidth

the page width (in inches) when building with pdflatex; passed to st_to_standalone(); see details.

border

passed as an option to standalone latex output type; see details.

ntex

number of times to build the pdf file

ltversion

numeric version number for the longtable package; newer versions have an issue that will break this code for longtables; so we are requiring an older version at this point.